Split Discord service routing tests into dispatch validation and flow-command routing specs
Extract fallback-sensitive tests from test_service_routing.py into two
dedicated files: test_dispatch_validation.py (17 tests) for dispatch ack
failures, scheduler serialization, and malformed payload handling; and
test_flow_command_routing.py (7 tests) for flow resume/restart/reply
picker fallbacks and pending state management.
